split JSX component into seperate files.
