+++ cypress/e2e/widget/masonry.cy.js
@@ -1,14 +1,11 @@
-describe('Should test the masonry', () => {
-    it('Should pass default tests', () => {
-        cy.widgetTests('masonry');
-    })
+describe("Should test the masonry", () => {
+  it("Should pass default tests", () => {
+    cy.widgetTests("masonry")
+  })
 
-    it('Should display timephrase in the tile', () => {
-        cy.visitWidget('masonry');
+  it("Should display timephrase in the tile", () => {
+    cy.visitWidget("masonry")
 
-        cy.getFirstTile().find('.tile-timephrase')
-        .should('exist')
-        .invoke('text')
-        .should('not.be.empty');
-    })
+    cy.getFirstTile().find(".tile-timephrase").should("exist").invoke("text").should("not.be.empty")
+  })
 })
