+++ cypress/e2e/widget/nightfall.cy.js
@@ -9,12 +9,9 @@ describe("Should test the nightfall template", () => {
     cy.getFirstTile().find(".tile-timephrase").should("exist").invoke("text").should("not.be.empty")
   })
 
-  it('Shoud display caption in the tile', () => {
-    cy.visitWidget('nightfall');
+  it("Shoud display caption in the tile", () => {
+    cy.visitWidget("nightfall")
 
-    cy.getFirstTile().find('.tile-caption')
-    .should('exist')
-    .invoke('text')
-    .should('not.be.empty');
-})
+    cy.getFirstTile().find(".tile-caption").should("exist").invoke("text").should("not.be.empty")
+  })
 })
