+++ cypress/e2e/widget/slider.cy.js
@@ -1,14 +1,11 @@
-describe('Should test the slider', () => {
-    it('Should pass default tests', () => {
-        cy.widgetTests('slider');
-    })
+describe("Should test the slider", () => {
+  it("Should pass default tests", () => {
+    cy.widgetTests("slider")
+  })
 
-    it('Should display timephrase in the tile', () => {
-        cy.visitWidget('slider');
+  it("Should display timephrase in the tile", () => {
+    cy.visitWidget("slider")
 
-        cy.getFirstTile().find('.tile-timephrase')
-        .should('exist')
-        .invoke('text')
-        .should('not.be.empty');
-    })
+    cy.getFirstTile().find(".tile-timephrase").should("exist").invoke("text").should("not.be.empty")
+  })
 })
