+++ widgets/quadrant/components/exit.button.tsx
@@ -1,0 +1,9 @@
+import { createElement } from "jsx-html"
+
+export const ExitButton = () => {
+  return (
+    <a className="exit" href="#">
+      <span className="widget-icon close-white"></span>
+    </a>
+  )
+}
